Interface injection + package move + requested changes + bug fixes
