+++ library/item/bows/src/main/java/org/quiltmc/qsl/item/bows/api/ExtendedCrossbowItem.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.bows.api;
+package org.quiltmc.qsl.item.bows.api;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -24,10 +24,10 @@ import net.minecraft.item.CrossbowItem;
 import net.minecraft.item.ItemStack;
 
 /**
- * This is the default implementation for {@link CrossbowExtensions}, allowing for the easy creation of new crossbows with no new modded functionality. <br>
+ * This is the default implementation for {@link CrossbowExtensions}, allowing for the easy creation of new crossbows with no new modded functionality. <p>
  * In order to have this crossbow edit the properties of shot projectiles, you must call {@code ShotProjectileEvents.CROSSBOW_MODIFY_SHOT_PROJECTILE.register(this);} for it to call {@link ExtendedCrossbowItem#onProjectileShot(ItemStack, ItemStack, LivingEntity, PersistentProjectileEntity)}
  */
-public class ExtendedCrossbowItem extends CrossbowItem implements CrossbowExtensions {
+public class ExtendedCrossbowItem extends CrossbowItem implements CrossbowExtensions, ShotProjectileEvents.ModifyProjectileFromCrossbow {
 	public ExtendedCrossbowItem(Settings settings) {
 		super(settings);
 	}
