+++ library/item/bows/src/main/java/org/quiltmc/qsl/item/bows/mixin/BowAttackMixin.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.bows.mixin;
+package org.quiltmc.qsl.item.bows.mixin;
 
-import org.quiltmc.qsl.bows.api.ShotProjectileEvents;
+import org.quiltmc.qsl.item.bows.api.ShotProjectileEvents;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Redirect;
