+++ library/item/bows/src/main/java/org/quiltmc/qsl/item/bows/mixin/EntityUseBowMixin.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.quiltmc.qsl.bows.mixin;
+package org.quiltmc.qsl.item.bows.mixin;
 
-import org.quiltmc.qsl.bows.api.BowExtensions;
+import org.quiltmc.qsl.item.bows.api.BowExtensions;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
